:bomb: moved magic into FSM
